Move naming code around in prep for replacement
